Move primitive type encoding to proto
