+++ Sources/BinaryCodable/Primitives/Bool+Coding.swift
@@ -20,10 +20,3 @@ extension Bool: DecodablePrimitive {
         self = data[data.startIndex] > 0
     }
 }
-
-extension Bool: ProtobufCodable {
-    
-    var protoType: String { "bool" }
-
-    static var zero: Bool { false }
-}
