+++ Sources/BinaryCodable/Primitives/Float+Coding.swift
@@ -21,16 +21,3 @@ extension Float: DecodablePrimitive {
         self.init(bitPattern: value)
     }
 }
-
-extension Float: ProtobufCodable {
-
-    func protobufData() throws -> Data {
-        data().swapped
-    }
-
-    init(fromProtobuf data: Data, path: [CodingKey]) throws {
-        try self.init(decodeFrom: data.swapped, path: path)
-    }
-
-    var protoType: String { "float" }
-}
