+++ Sources/BinaryCodable/Primitives/Int+Coding.swift
@@ -33,29 +33,6 @@ extension Int: VariableLengthCodable {
     }
 }
 
-extension Int: FixedSizeCompatible {
-
-    public static var fixedSizeDataType: DataType {
-        .eightBytes
-    }
-
-    public var fixedProtoType: String {
-        "sfixed64"
-    }
-
-    public init(fromFixedSize data: Data, path: [CodingKey]) throws {
-        let signed = try Int64(fromFixedSize: data, path: path)
-        guard let value = Int(exactly: signed) else {
-            throw DecodingError.variableLengthEncodedIntegerOutOfRange(path)
-        }
-        self = value
-    }
-
-    public var fixedSizeEncoded: Data {
-        Int64(self).fixedSizeEncoded
-    }
-}
-
 extension Int: SignedValueCompatible {
 
     public var positiveProtoType: String {
@@ -84,21 +61,5 @@ extension Int: ZigZagCodable {
             throw DecodingError.variableLengthEncodedIntegerOutOfRange(path)
         }
         self = value
-    }
-}
-
-extension Int: ProtobufEncodable {
-
-    func protobufData() -> Data {
-        variableLengthEncoding
-    }
-
-    var protoType: String { "sint64" }
-}
-
-extension Int: ProtobufDecodable {
-
-    init(fromProtobuf data: Data, path: [CodingKey]) throws {
-        try self.init(fromVarint: data, path: path)
     }
 }
