+++ Sources/BinaryCodable/Primitives/Int64+Coding.swift
@@ -65,50 +65,9 @@ extension Int64: ZigZagDecodable {
     }
 }
 
-extension Int64: FixedSizeCompatible {
-
-    static public var fixedSizeDataType: DataType {
-        .eightBytes
-    }
-
-    public var fixedProtoType: String {
-        "sfixed64"
-    }
-
-    public init(fromFixedSize data: Data, path: [CodingKey]) throws {
-        guard data.count == MemoryLayout<UInt64>.size else {
-            throw DecodingError.invalidDataSize(path)
-        }
-        let value = UInt64(littleEndian: read(data: data, into: UInt64.zero))
-        self.init(bitPattern: value)
-    }
-
-    public var fixedSizeEncoded: Data {
-        let value = UInt64(bitPattern: littleEndian)
-        return toData(value)
-    }
-}
-
 extension Int64: SignedValueCompatible {
 
     public var positiveProtoType: String {
         "int64"
     }
-}
-
-extension Int64: ProtobufEncodable {
-
-    func protobufData() -> Data {
-        variableLengthEncoding
-    }
-
-    var protoType: String { "sint64" }
-}
-
-extension Int64: ProtobufDecodable {
-
-    init(fromProtobuf data: Data, path: [CodingKey]) throws {
-        try self.init(fromVarint: data, path: path)
-    }
-
 }
