+++ Sources/BinaryCodable/Protobuf/Primitives/Bool+Protobuf.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+extension Bool: ProtobufCodable {
+
+    var protoType: String { "bool" }
+
+    static var zero: Bool { false }
+}
