+++ Sources/BinaryCodable/Protobuf/Primitives/Data+Protobuf.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+extension Data: ProtobufCodable {
+
+    var protoType: String { "bytes" }
+
+    static var zero: Data { .empty }
+}
