+++ Sources/BinaryCodable/Protobuf/Primitives/Int64+Protobuf.swift
@@ -1,0 +1,18 @@
+import Foundation
+
+extension Int64: ProtobufEncodable {
+
+    func protobufData() -> Data {
+        variableLengthEncoding
+    }
+
+    var protoType: String { "sint64" }
+}
+
+extension Int64: ProtobufDecodable {
+
+    init(fromProtobuf data: Data, path: [CodingKey]) throws {
+        try self.init(fromVarint: data, path: path)
+    }
+
+}
