+++ Sources/BinaryCodable/Protobuf/Primitives/String+Protobuf.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+extension String: ProtobufCodable {
+
+    var protoType: String { "string" }
+
+    static let zero = ""
+}
