+++ Sources/BinaryCodable/Wrappers/FixedSizeCompatible.swift
@@ -3,12 +3,6 @@ import Foundation
 /// An integer type which can be forced to use a fixed-length encoding instead of variable-length encoding.
 public protocol FixedSizeCompatible {
 
-    /// The wire type of the type, which has a constant length
-    static var fixedSizeDataType: DataType { get }
-
-    /// The protobuf type equivalent to the fixed size type
-    var fixedProtoType: String { get }
-
     /// The value encoded as fixed size binary data
     var fixedSizeEncoded: Data { get }
 
@@ -18,12 +12,4 @@ public protocol FixedSizeCompatible {
      - Throws: `DecodingError`
      */
     init(fromFixedSize data: Data, path: [CodingKey]) throws
-}
-
-extension FixedSizeCompatible {
-
-    /// The wire type of the value, which has a constant length
-    var fixedSizeDataType: DataType {
-        Self.fixedSizeDataType
-    }
 }
